git-svn-id: svn://192.168.1.10/public/CoffeeMud@14371 0d6f1817-ed0e-0410-87c9-987e46238f29

+++ com/planet_ink/coffee_mud/Abilities/Properties/Prop_HaveAdjuster.java
@@ -728,7 +728,7 @@ public class Prop_HaveAdjuster extends Property implements TriggeredAffect
 						case PhyStats.STAT_HEIGHT:
 							break;
 						case PhyStats.NUM_STATS:
-							level+= (((Integer) changes[c + 1]).intValue() * 5);
+							level+= (((Double) changes[c + 1]).intValue() * 5);
 							break;
 						}
 					}
